# Optimize GraphHelper#resources_by_date