@efatsi 2016-05-26
that's some dense code, but good work. if this was something we were actively developing I'd probably suggest cleaning it up, but 👍