@mackermedia 2016-05-26
👍  Cool to see the pre & post optimization benchmarks. Nice work!